bot/requirements.txt
python-telegram-bot
SQLAlchemy
boto3
psycopg2-binary
Pillow